requirements.txt
Django==5.2.4
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
pyodbc==5.1.0
django-mssql-backend==2.8.1
